회원가입 api 비동기 방식으로 수정 중

+++ src/routers/comment.js
@@ -6,6 +6,8 @@ const conn = require('../../database/connect/maria');
 //Apis
 //댓글 업로드
 router.post("/:postidx", (req, res) => {
+    //postidx가 매칭이 안 되니까 바디로 보내주는게 맞다.
+    //crud querystring으로 빼거나.
     const postIdx = req.params.postidx;
     const {content} = req.body;
     const uploadCommentResult = {
@@ -18,15 +20,11 @@ router.post("/:postidx", (req, res) => {
         console.log(idx, postIdx, content)
         //db에 값 입력하기
         conn.query('INSERT INTO comment (post_idx, account_idx, content) VALUES (?, ?, ?)', [postIdx, idx, content], (err) => {
-            if (err) {
-                throw new Error("데이터베이스가 이상해요")
-            } 
-            else {
-                console.log("성공");
-                uploadCommentResult.success = true;
-                uploadCommentResult.message = "댓글이 업로드되었습니다.";
-                res.send(uploadCommentResult)
-            }
+            if (err) throw new Error("데이터베이스가 이상해요");
+            console.log("성공");
+            uploadCommentResult.success = true;
+            uploadCommentResult.message = "댓글이 업로드되었습니다.";
+            res.send(uploadCommentResult)
         });
     }
     catch (e) {
@@ -42,12 +40,23 @@ router.get("/:postidx/comment", (req, res) => {
     const perPage = req.query.per_page || 10;
     const viewCommentResult = {
         "success": false,
-        "message": ""
+        "message": "",
+        "comments": []
     }
     try {
         if (!req.session.user) throw new Error("세션에 사용자 정보가 없습니다.");
-        
-        conn.query('SELECT * from comment')
+    
+        conn.query('SELECT account_idx, content from comment WHERE post_idx=?', [postIdx], (err) => {
+            if (err) {
+                throw new Error("데이터베이스가 이상해요");
+            } 
+            else {
+                viewCommentResult.success = true;
+                viewCommentResult.message = "댓글 불러오기 성공";
+                viewCommentResult.comments = results; 
+                res.send(viewCommentResult)
+            }
+        });
     }
     catch (e) {
         viewCommentResult.message = e.message;
@@ -69,7 +78,7 @@ router.put("/:commentidx", (req, res) => {
         const idx = req.session.user.idx;
 
         //db에 값 입력하기
-        conn.query('UPDATE comment SET content=? WHERE idx=?', [content, contentIdx], (err) => {
+        conn.query('UPDATE comment SET content=? WHERE idx=? AND account_idx=?', [content, contentIdx, idx], (err) => {
             if (err) {
                 throw new Error("데이터베이스가 이상해요")
             } 
@@ -97,9 +106,9 @@ router.delete("/:commentidx", (req, res) => {
     try {
         if (!req.session.user) throw new Error("세션에 사용자 정보가 없습니다.");
         const idx = req.session.user.idx;
-
+        
         //db에 값 입력하기
-        conn.query('DELETE FROM comment WHERE idx=?', [contentIdx], (err) => {
+        conn.query('DELETE FROM comment WHERE idx=? AND account_idx=?', [contentIdx, idx], (err) => {
             if (err) {
                 throw new Error("데이터베이스가 이상해요")
             }
